+++ oganesson/utilities/atomic_data.py
@@ -2,7 +2,7 @@
 s_groups = [[
     4, 12, 20, 38, 56
 ],
-    [3, 11, 19, 37, 55]
+    [1, 3, 11, 19, 37, 55]
 ]
 
 rare_earths = [x for x in range(57, 72)] + [x for x in range(89, 95)]
@@ -18,12 +18,16 @@ d_groups_no_REAs = [
     [21, 39]
 ]
 
-d_groups_flat = [item for col in d_groups for item in col]
-
 d_groups_no_REAs += [[x, x+18, x+50] for x in range(22, 31)]
 
-p_groups = [[x, x+8, x+26, x+44, x+76] for x in range(5, 8)]
+p_groups = [[2]]
+p_groups += [[x, x+8, x+26, x+44, x+76] for x in range(5, 8)]
 p_groups += [[x, x+8, x+26, x+44] for x in range(8, 11)]
+
+p_groups_flat = [item for col in p_groups for item in col]
+s_groups_flat = [item for col in s_groups for item in col]
+d_groups_flat = [item for col in d_groups for item in col]
+
 
 all_groups = d_groups + p_groups + s_groups
 all_groups_no_REAs = d_groups_no_REAs + p_groups + s_groups
@@ -48,4 +52,13 @@ def env_friendly(r):
     else:
         return False
   
-  
+def get_group(a):
+    if a in d_groups_flat:
+        return "d"
+    elif a in s_groups_flat:
+        return "s"
+    elif a in p_groups_flat:
+        return "p"
+    else:
+        print('Atomic number:',a)
+        raise Exception('Wrong atomic number!')